Restructure most of the code.
