+++ electron/constants.ts
@@ -14,6 +14,7 @@ const legendaryBin = fixPathForAsarUnpack(join(__dirname, '/bin/legendary'))
 const icon = fixPathForAsarUnpack(join(__dirname, '/icon.png'))
 const iconDark = fixPathForAsarUnpack(join(__dirname, '/icon-dark.png'))
 const iconLight = fixPathForAsarUnpack(join(__dirname, '/icon-light.png'))
+const installed = `${legendaryConfigPath}/installed.json`
 const loginUrl =
   'https://www.epicgames.com/id/login?redirectUrl=https%3A%2F%2Fwww.epicgames.com%2Fid%2Fapi%2Fredirect'
 const sidInfoUrl =
@@ -37,6 +38,7 @@ export {
   icon,
   iconDark,
   iconLight,
+  installed,
   legendaryBin,
   legendaryConfigPath,
   loginUrl,
