+++ electron/legendary_utils/legacy_library.ts
@@ -1,73 +1,55 @@
-import { exec } from 'child_process'
 import { existsSync, readFileSync, readdirSync, stat } from 'graceful-fs'
-import { promisify } from 'util'
 import prettyBytes from 'pretty-bytes'
+/* eslint-disable @typescript-eslint/ban-ts-comment */
+import { promisify } from 'util'
 
-import { Game, InstalledInfo, KeyImage } from './types'
-import { GameInfo } from './game'
-import { legendaryBin, legendaryConfigPath } from './constants'
+import { InstalledInfo, KeyImage, RawGameJSON, UserInfo } from '../types'
+import { getUserInfo, isLoggedIn, writeDefaultConfig } from '../config'
+import { heroicConfigPath, legendaryConfigPath } from '../constants'
 
-const execAsync = promisify(exec)
 const statAsync = promisify(stat)
+const dlcs: string[] = []
 
-type LoadState = 'loaded' | 'loading' | 'not loaded'
+const installed = `${legendaryConfigPath}/installed.json`
 
-class Library {
-  private load_state: LoadState
-  private static globalInstance: Library = null
+export async function getLegendaryConfig(file: string): Promise<unknown> {
+  const loggedIn = isLoggedIn()
 
-  private library: Map<string, GameInfo>
+  if (!isLoggedIn) {
+    return { library: [], user: { displayName: null } }
+  }
 
-  private constructor(lazy_load: boolean) {
-    this.load_state = 'not loaded'
-    if (lazy_load) {
-      this.load()
+  const files: {
+    config: string
+    installed: RawGameJSON[]
+    library: string
+    user: UserInfo
+  } = {
+    config: heroicConfigPath,
+    installed: await statAsync(installed)
+      .then(() => JSON.parse(readFileSync(installed, 'utf-8')))
+      .catch(() => []),
+    library: `${legendaryConfigPath}/metadata/`,
+    user: getUserInfo()
+  }
+
+  if (file === 'user') {
+    if (loggedIn) {
+      await writeDefaultConfig()
+      return files.user.displayName
     }
+    return null
   }
 
-  public static get(lazy_load = true) {
-    if (this.globalInstance === null) {
-      Library.globalInstance = new Library(lazy_load)
-    }
-    return this.globalInstance
-  }
-
-  public async getGames() {
-    return Array.from(this.library.values()).sort(
-      (a: { title: string }, b: { title: string }) => {
-        const gameA = a.title.toUpperCase()
-        const gameB = b.title.toUpperCase()
-        return gameA < gameB ? -1 : 1
-      }
-    )
-  }
-
-  public async getGameInfo(app_name: string) {
-    return this.library.get(app_name) || null
-  }
-
-  public async listUpdateableGames() {
-    const command = `${legendaryBin} list-installed --check-updates --tsv | grep True | awk '{print $1}'`
-    const { stdout } = await execAsync(command)
-    const result = stdout.split('\n')
-    return result
-  }
-
-  private async load() {
-    if (this.load_state == 'loading') {
-      // there is a thread already loading stuff, wait for it to finish.
-    }
-    this.load_state = 'loading'
-
-    const libraryPath = `${legendaryConfigPath}/metadata/`
+  if (file === 'library') {
     const fallBackImage =
       'https://user-images.githubusercontent.com/26871415/103480183-1fb00680-4dd3-11eb-9171-d8c4cc601fba.jpg'
 
-    if (existsSync(libraryPath)) {
-      return readdirSync(libraryPath)
-        .map((file) => `${libraryPath}/${file}`)
+    if (existsSync(files.library)) {
+      return readdirSync(files.library)
+        .map((file) => `${files.library}/${file}`)
         .map((file) => JSON.parse(readFileSync(file, 'utf-8')))
-        .forEach(({ app_name, metadata, asset_info }) => {
+        .map(({ app_name, metadata, asset_info }) => {
           const {
             description,
             keyImages,
@@ -78,8 +60,6 @@ class Library {
           } = metadata
 
           const { namespace } = asset_info
-
-          const dlcs: string[] = []
 
           if (dlcItemList) {
             dlcItemList.forEach(
@@ -108,12 +88,7 @@ class Library {
           const art_logo = logo ? logo.url : null
           const art_square = gameBoxTall ? gameBoxTall.url : fallBackImage
 
-          const installedJSON = `${legendaryConfigPath}/installed.json`
-          const installedGames: Game[] = Object.values(
-            statAsync(installedJSON)
-              .then(() => JSON.parse(readFileSync(installedJSON, 'utf-8')))
-              .catch(() => [])
-          )
+          const installedGames: RawGameJSON[] = Object.values(files.installed)
 
           const isInstalled = Boolean(
             installedGames.filter((game) => game.app_name === app_name).length
@@ -134,7 +109,7 @@ class Library {
           const convertedSize =
             install_size && prettyBytes(Number(install_size))
 
-          this.library.set(app_name, {
+          return {
             app_name,
             art_cover: art_cover || art_square,
             art_logo,
@@ -143,7 +118,6 @@ class Library {
             description,
             developer,
             executable,
-            extraInfo: [],
             folderName: installFolder,
             info,
             install_path,
@@ -154,10 +128,14 @@ class Library {
             saveFolder,
             title,
             version
-          } as GameInfo)
+          }
+        })
+        .sort((a: { title: string }, b: { title: string }) => {
+          const gameA = a.title.toUpperCase()
+          const gameB = b.title.toUpperCase()
+          return gameA < gameB ? -1 : 1
         })
     }
+    return { library: [], user: null }
   }
 }
-
-export { Library }
